Argument validation

check_type checks if argument has correct class

check_function checks if argument is a function

check_numeric checks if argument is numeric

check_positive checks if argument is positive

check_non_negative checks if argument is non-negative

check_positive_all checks if argument has only positive values

check_non_negative_all checks if argument has only non-negative values

check_integer_all checks if argument has only integer values

check_interval checks if argument is inside a given interval

check_not_null checks if argument is not null

check_false checks if argument is false or zero

check_length checks if argument has given length

check_length_1_or checks if argument has given length or length 1

check_length_geq checks if argument has at least a given length

check_lengths checks if two arguments have equal length

check_in_data checks that variable exists in a data frame

check_all_leq checks that argument has values less than equal to given maximums, elementwise

check_not_named checks that names(arg) is NULL

check_named checks that length of names(arg) is greater than zero

check_dim checks that arg has expected number of dimensions

check_null checks that arg is NULL

check_type(arg, allowed)

check_function(arg, arg_name)

check_numeric(arg)

check_positive(arg)

check_non_negative(arg)

check_positive_all(arg)

check_non_negative_all(arg)

check_integer_all(arg)

check_interval(arg, lower, upper)

check_not_null(arg)

check_null(arg, reason = NULL)

check_length(arg, expected)

check_length_1_or(arg, expected)

check_length_geq(arg, min_length)

check_lengths(a, b)

check_false(arg)

check_in_data(var_name, data)

check_all_leq(arg, maximums)

check_not_named(arg)

check_named(arg)

check_dim(arg, D)

Arguments

arg
An argument to check.

allowed
Allowed class names.

arg_name
argument name

lower
interval minimum

upper
interval maximum

reason
explanation why the argument should be NULL

expected
Expected length.

min_length
Minimum allowed length.

a
An argument to check.

b
An argument to check.

var_name
the variable to be searched for

data
a data frame

maximums
maximum allowed values for arg

D
expected number of dimensions

Value

TRUE if the check passes.

See also