Kernel matrix computation using the C++ functions of the package namespace

Kernel matrices using parameter values defined by each posterior draw can be computed using kernel_matrices. It returns a 4-dimensional array of shape num_draws x num_comps x n1 x n2.

The constant kernel matrices that do not depend on parameters can be computed using kernel_matrices_const. It returns a list of kernel matrices with length equal to number of model components.

Input information can be translated from a data frame to the needed format using kernel_matrices_create_input.

kernel_matrices(model, x1, x2, theta, verbose, STREAM = get_stream())

kernel_matrices_const(model, x1, x2, STREAM = get_stream())

kernel_matrices_create_input(model, x)

Arguments

model
An object of class lgpmodel.

x1
A list defining the first input argument. Can be created by kernel_matrices_create_input.

x2
A list defining the second input argument. Can be created by kernel_matrices_create_input.

theta
A list of kernel parameter sets.

verbose
Should more information and a progress bar be printed?

STREAM
External pointer. By default obtained with get_stream.

x
A data frame of points where the predictions are computed. The function new_x provides an easy way to create it.

See also