An S4 class to represent the output of the lgp function

An S4 class to represent the output of the lgp function

# S4 method for lgpfit
show(object)

# S4 method for lgpfit
component_names(object)

# S4 method for lgpfit
num_components(object)

# S4 method for lgpfit
postproc(object, verbose = TRUE)

# S4 method for lgpfit
contains_postproc(object)

# S4 method for lgpfit
clear_postproc(object)

# S4 method for lgpfit
get_model(object)

# S4 method for lgpfit
get_stanfit(object)

# S4 method for lgpfit
is_f_sampled(object)

# S4 method for lgpfit,missing
plot(x, y)

Arguments

object
The object for which to apply a class method.

verbose
Can the method print any messages?

x
an lgpfit object to visualize

y
unused argument

Methods (by generic)

show: Print information and summary about the fit object.

component_names: Get names of model components.

num_components: Get number of model components. Returns a positive integer.

postproc: Apply postprocessing. Returns an updated lgpfit object (copies data).

contains_postproc: Check if object contains postprocessing information.

clear_postproc: Returns an updated (copies data) lgpfit object without any postprocessing information.

get_model: Get the stored lgpmodel object. Various properties of the returned object can be accessed as explained in the documentation of lgpmodel.

get_stanfit: Get the stored stanfit object. Various properties of the returned object can be accessed or plotted as explained here or in the documentation of stanfit.

is_f_sampled: Determine if inference was done by sampling the latent signal f (and its components).

plot: Visualize parameter draws using plot_draws.

Slots

stan_fit

An object of class stanfit.

model

An object of class lgpmodel.

num_draws

Total number of parameter draws.

postproc_results

A named list containing possible postprocessing results.

See also