An S4 class to represent an additive GP model

An S4 class to represent an additive GP model

# S4 method for lgpmodel
show(object)

# S4 method for lgpmodel
parameter_info(object, digits = 3)

# S4 method for lgpmodel
component_info(object)

# S4 method for lgpmodel
covariate_info(object)

# S4 method for lgpmodel
component_names(object)

# S4 method for lgpmodel
is_f_sampled(object)

Arguments

object
The object for which to apply a class method.

digits
number of digits to show for floating point numbers

Methods (by generic)

show: Print information and summary about the object. Returns object invisibly.

parameter_info: Get a parameter summary (bounds and priors). Returns a data.frame.

component_info: Get a data frame with information about each model component.

covariate_info: Get covariate information.

component_names: Get names of model components.

is_f_sampled: Determine if inference of the model requires sampling the latent signal f (and its components).

Slots

formula

An object of class lgpformula

data

The original unmodified data.

stan_input

The data to be given as input to rstan::sampling

var_names

List of variable names grouped by type.

var_scalings

A named list with fields

y - Response variable normalization function and its inverse operation. Must be an lgpscaling object.

x_cont - Continuous covariate normalization functions and their inverse operations. Must be a named list with each element is an lgpscaling object.

var_info

A named list with fields

x_cat_levels - Names of the levels of categorical covariates before converting from factor to numeric.

info

Other info in text format.

sample_f

Whether the signal f is sampled or marginalized.

full_prior

Complete prior information.