Local Gaussian Prediction for Regression Uncertainty Modelling

Local Gaussian Prediction for Regression Uncertainty Modelling

lgpr(location, scale, interval, n = 512, alpha = 0.05)

Arguments

location
a numeric vector for defining the location of the gaussian distribution. See Details.

scale
a numeric vector for defining the location of the gaussian distribution. See Details.

interval
a vector with two numeric values defining the upper and lower bound of the discrete gaussian distribution.

n
number of uniform steps defining the values at which the density of the distribution (gaussian) is evaluated. Default is 512.

alpha
tolerance of the confidence interval created by the gaussian distribution. confidence = 1 - alpha.

Value

a lgpr object

Details

The location argument of the gaussian distribution in this context refers to the point estimate or prediction obtained from a regression algorithm and the scale argument refers to an estimate of the variance (error) associated to such point estimate. If the regression algorithm do not output a variance estimate by default, the variance can be model from the residuals of the model for instance is up to the user / analyst to come up with a strategy to model the variance (scale) associated with the regression predictions (location). This function is intended to ease up the computation of prediction interval when assuming a gaussian distribution for the residuals of the regression model is reasonable. If the scale argument is composed of different values the predictive intervals generated will be heteroscedastic. That is the reason why this approach is named Local Gaussian Prediction because the predictive interval are adjusted to difficulty of the prediction itself through the scale argument. In this implementation it is assumed that the prediction domain is bounded and the limits are given by the interval argument. The prediction interval is computed from the truncated gaussian distribution by a discretization approximation.

Author

David Senhora Navega