Plot a generated/fit model component

Data frames specified in arguments df, and df_err must have a format where

The first column is the grouping factor (usually id).

The second column is the x-axis variable (usually age).

The third column is the coloring factor. If name of the third column is NA, coloring is not done.

A column named y must contain the y-axis variable (not for df_err).

A column named lower (upper) must contain the lower (upper) bound of error bar (only for df_err).

The posterior draw using which the fit has been computed can be specified with a factor named _draw_ (only for df).

plot_api_c(
  df,
  df_err = NULL,
  alpha = 1,
  alpha_err = 0.2,
  no_err = FALSE,
  no_line = FALSE
)

Arguments

df
a data frame

df_err
a data frame

alpha
line opacity

alpha_err
ribbon opacity

no_err
hide error bar even when it would normally be plotted?

no_line
hide line even when it would normally be plotted?

Value

A ggplot object.

See also