Visualize all model components

This calls plot_f for all model components.

plot_components(
  fit,
  x = NULL,
  pred = NULL,
  group_by = "id",
  t_name = "age",
  draws = NULL,
  reduce = mean,
  MULT_STD = 2,
  color_by = NA,
  no_err = FALSE,
  ylim = NULL,
  draw = TRUE,
  nrow = NULL,
  ncol = NULL,
  gg_add = NULL,
  ...
)

Arguments

fit
An object of class lgpfit.

x
A data frame of prediction points, containing at least the variables used as covariates in the model. This must be specified if pred is not NULL.

pred
An object of class GaussianPrediction or Prediction.

group_by
name of the grouping variable (use group_by=NA to avoid grouping)

t_name
name of the x-axis variable

draws
only has effect if pred is NULL

reduce
only has effect if pred is NULL

MULT_STD
a multiplier for standard deviation

color_by
Names of coloring factors. Can have length 1 or equal to the number of components. See the color_by argument of plot_f.

no_err
Should the error ribbons be skipped even though they otherwise would be shown? Can have length 1 or equal to number of components + 1. See the no_err argument of plot_api_c.

ylim
a vector of length 2 (upper and lower y-axis limits), or NULL

draw
if this is TRUE, the plot grid is drawn using arrangeGrob

nrow
number of grid rows

ncol
number of grid columns

gg_add
additional ggplot obejct to add to each plot

...
additional arguments to plot_api_c

Value

a list of ggplot objects invisibly

See also