Visualize the distribution of parameter draws

Visualize the distribution of parameter draws

plot_draws(
  fit,
  type = "intervals",
  regex_pars = c("alpha", "ell", "wrp", "sigma", "phi", "gamma"),
  ...
)

plot_beta(fit, type = "dens", verbose = TRUE, ...)

plot_warp(
  fit,
  num_points = 300,
  window_size = 48,
  color = colorset("red", "dark"),
  alpha = 0.5
)

plot_effect_times(fit, type = "areas", verbose = TRUE, ...)

Arguments

fit

an object of class lgpfit

type

plot type, allowed options are "intervals", "dens", "areas", and "trace"

regex_pars

regex for parameter names to plot

...

additional arguments for the bayesplot function mcmc_intervals, mcmc_dens, mcmc_areas or mcmc_trace

verbose

Can any output be printed?

num_points

number of plot points

window_size

width of time window

color

line color

alpha

line alpha

Value

a ggplot object or list of them

Functions

plot_draws(): visualizes the distribution of any set of model parameters (defaults to kernel hyperparameters and possible observation model parameters)

plot_beta(): visualizes the distribution of the individual-specific disease effect magnitude parameter draws

plot_warp(): visualizes the input warping function for different draws of the warping steepness parameter

plot_effect_times(): visualizes the input warping function for different parameter draws

See also