Compute analytic out-of-sample predictions

Compute analytic out-of-sample predictions

pred.gaussian_compute(kernels, y, sigma, delta, verbose, STREAM = get_stream())

Arguments

kernels
A list returned by pred_kernels.

y
response variable vector of length num_obs

sigma
a vector with length equal to number of parameter sets

delta
jitter to ensure positive definite matrices

verbose
Should more information and a progress bar be printed?

STREAM
External pointer. By default obtained with get_stream.

Value

A list.

See also