Compute all kernel matrices required for computing predictions

Compute all kernel matrices required for computing predictions

pred_kernels(fit, x, reduce, draws, verbose, STREAM = get_stream())

Arguments

fit
An object of class lgpfit.

x
A data frame of points where the predictions are computed. The function new_x provides an easy way to create it.

reduce
Reduction for parameters draws. Can be a function that is applied to reduce all parameter draws into one parameter set, or NULL (no reduction). Has no effect if draws is specified.

draws
Indices of parameter draws to use, or NULL to use all draws.

verbose
Should more information and a progress bar be printed?

STREAM
External pointer. By default obtained with get_stream.

Value

Returns a list with fields named data_vs_data, pred_vs_data and pred_vs_pred. Each of them is a list with length equal to the number of parameter sets.

See also