Prior (predictive) sampling

These functions take an lgpmodel object, and

prior_pred samples from the prior predictive distribution of the model

sample_param_prior samples only its parameter prior using sampling

prior_pred(
  model,
  verbose = TRUE,
  quiet = FALSE,
  refresh = 0,
  STREAM = get_stream(),
  ...
)

sample_param_prior(model, verbose = TRUE, quiet = FALSE, ...)

Arguments

model
An object of class lgpmodel.

verbose
Should more information and a possible progress bar be printed?

quiet
This forces verbose to be FALSE. If you want to suppress also the output from Stan, give the additional argument refresh=0.

refresh
Argument for sampling.

STREAM
External pointer. By default obtained with rstan::get_stream().

...
Additional arguments for sampling.

Value

prior_pred returns a list with components

y_draws: A matrix containing the prior predictive draws as rows. Can be passed to bayesplot::pp_check() for graphical prior predictive checking.

pred_draws: an object of class Prediction, containing prior draws of each model component and their sum

param_draws: a stanfit object of prior parameter draws (obtained by calling sample_param_prior internally)

sample_param_prior returns an object of class stanfit

See also