Fitting a model

sample_model takes an lgpmodel object and fits it using sampling.

optimize_model takes an lgpmodel object and fits it using optimizing.

sample_model(
  model,
  verbose = TRUE,
  quiet = FALSE,
  skip_postproc = is_f_sampled(model),
  ...
)

optimize_model(model, ...)

Arguments

model
An object of class lgpmodel.

verbose
Can messages be printed?

quiet
Should all output messages be suppressed? You need to set also refresh=0 if you want to suppress also the progress update messages from sampling.

skip_postproc
Should all postprocessing be skipped? If this is TRUE, the returned lgpfit object will likely be much smaller (if sample_f=FALSE).

...
Optional arguments passed to sampling or optimizing.

Value

sample_model returns an object of class lgpfit containing the parameter draws, the original model object, and possible postprocessing results. See documentation of lgpfit for more information.

optimize_model directly returns the list returned by optimizing. See its documentation for more information.

See also