Select relevant components

select performs strict selection, returning either TRUE or FALSE for each component.

select.integrate is like select, but instead of a fixed threshold, computes probabilistic selection by integrating over a threshold density.

select_freq performs the selection separately using each parameter draw and returns the frequency at which each component was selected.

select_freq.integrate is like select_freq, but instead of a fixed threshold, computes probabilistic selection frequencies by integrating over a threshold density.

select(fit, reduce = function(x) base::mean(x), threshold = 0.95, ...)

select_freq(fit, threshold = 0.95, ...)

select.integrate(
  fit,
  reduce = function(x) base::mean(x),
  p = function(x) stats::dbeta(x, 100, 5),
  h = 0.01,
  verbose = TRUE,
  ...
)

select_freq.integrate(
  fit,
  p = function(x) stats::dbeta(x, 100, 5),
  h = 0.01,
  verbose = TRUE,
  ...
)

Arguments

fit
An object of class lgpfit.

reduce
The reduce argument for relevances.

threshold
Threshold for relevance sum. Must be a value between 0 and 1.

...
Additional arguments to relevances.

p
A threshold density over interval [0,1].

h
A discretization parameter for computing a quadrature.

verbose
Should this show a progress bar?

Value

See description.