Create covariate data for Stan input

Creates the following Stan data input list fields:

num_cov_cat

num_cov_cont

x_cat

x_cat_num_levels

x_cont

x_cont_unnorm

x_cont_mask

stan_data_covariates(data, x_names, x_cont_scl)

Arguments

data
a data frame

x_names
unique covariate names

x_cont_scl
Information on how to scale the continuous covariates. This can either be

an existing list of objects with class lgpscaling, or

NA, in which case such list is created by computing mean and standard deviation from data

Value

a named list with fields

to_stan: a list of stan data

x_cont_scaling: normalization function and inverse for each continuous covariate

x_cat_levels: names of the levels of each categorical covariate before conversion from factor to numeric