Get names of all variables appearing in a term

Get names of all variables appearing in a term

term_variables(term)

Arguments

term
an object of class lgpterm

Value

a list of variable names

See also