The 'lgpr' package.

A package for Bayesian additive Gaussian process (GP) modeling with interpretable kernels designed for longitudinal data.

Features inference of covariate effects and covariate relevance assesment.

Models can include non-stationary, heterogeneous and temporally uncertain effects.

Bayesian inference of the model (hyper)parameters using rstan.

Functions for visualizing longitudinal data, posterior draws, model predictions and inferred covariate effects are also provided.

Overview

See the documentation of

lgp for info on how to specify and fit models

ppc for prior and posterior predictive checks

relevances for component/covariate relevance assessment

select for component/covariate selection

get_pred for model predictions and inferred components at data points

pred for computing out-of-sample predictions and inferred components

plot_pred and plot_f for visualizing predictions and inferred components

Tutorials

See tutorials at https://jtimonen.github.io/lgpr-usage/index.html.

Citation

Run citation("lgpr") to get citation information.

Feedback

Bug reports, PRs, enchancement ideas or user experiences in general are welcome and appreciated. Create an issue in Github or email the author.

References

Carpenter, B. et al. (2017). Stan: A probabilistic programming language. Journal of Statistical Software 76(1).

Jonah Gabry, Ben Goodrich and Martin Lysy (2019). rstantools: Tools for Developing R Packages Interfacing with 'Stan'. R package version 2.0.0.

Gabry, J. and Mahr, T. (2019). bayesplot: Plotting for Bayesian Models. R package version 1.7.0, http://mc-stan.org/bayesplot.

Stan Development Team (2019). RStan: the R interface to Stan. R package version 2.19.2. http://mc-stan.org/.

Author

Juho Timonen (first.last at iki.fi)