Visualize the distribution of the obtained parameter draws

plot_draws visualizes the distribution of any set of model parameters (defaults to kernel hyperparameters and possible observation model parameters)

plot_beta visualizes the distribution of the individual-specific disease effect magnitude parameter draws

plot_effect_times visualizes the input warping function for different parameter draws

plot_warp visualizes the input warping function for different draws of the warping steepness parameter

plot_draws(
  fit,
  type = "intervals",
  regex_pars = c("alpha", "ell", "wrp", "sigma", "phi", "gamma"),
  ...
)

plot_warp(fit, num_points = 300, window_size = 48, color_scheme = "brightblue")

plot_beta(fit, type = "dens", ...)

plot_effect_times(fit, type = "areas", ...)

Arguments

fit
an object of class lgpfit

type
plot type, allowed options are "intervals", "dens", "areas", and "trace"

regex_pars
regex for parameter names to plot

...
additional arguments for the bayesplot function mcmc_intervals, mcmc_dens, mcmc_areas or mcmc_trace

num_points
number of plot points

window_size
width of time window

color_scheme
name of bayesplot color scheme

Value

a ggplot object or list of them