Computing out-of-sample predictions

pred calls either pred.gaussian or pred.kr depending on whether fit is for a model that marginalized or sampled the latent signal f

pred.gaussian computes analytic posterior or prior predictive distribution and returns an object of class GaussianPrediction

pred.kr computes predictions using kernel regression for the samples of f and returns an object of class Prediction

These also compute the predicted (distribution of the) total signal f and its additive components. All these are computed for each parameter draw (defined by draws), or other parameter set (obtained by a reduction defined by reduce).

pred(
  fit,
  x,
  reduce = function(x) base::mean(x),
  draws = NULL,
  verbose = TRUE,
  STREAM = get_stream()
)

pred.gaussian(fit, x, reduce, draws, verbose, STREAM = get_stream())

pred.kr(fit, x, reduce, draws, verbose, STREAM = get_stream())

Arguments

fit
An object of class lgpfit.

x
A data frame of points where the predictions are computed. The function new_x provides an easy way to create it.

reduce
Reduction for parameters draws. Can be a function that is applied to reduce all parameter draws into one parameter set, or NULL (no reduction). Has no effect if draws is specified.

draws
Indices of parameter draws to use, or NULL to use all draws.

verbose
Should more information and a progress bar be printed?

STREAM
External pointer. By default obtained with get_stream.

Value

An object of class GaussianPrediction or Prediction.

See also