S4 generics for lgpfit, lgpmodel, and other objects

S4 generics for lgpfit, lgpmodel, and other objects

parameter_info(object, digits)

component_info(object)

covariate_info(object)

component_names(object)

get_model(object)

is_f_sampled(object)

get_stanfit(object)

postproc(object, ...)

contains_postproc(object)

clear_postproc(object)

num_paramsets(object)

num_evalpoints(object)

num_components(object)

Arguments

object
object for which to apply the generic

digits
number of digits to show

...
additional optional arguments to pass

Value

parameter_info returns a data frame with one row for each parameter and columns for parameter name, parameter bounds, and the assigned prior

component_info returns a data frame with one row for each model component, and columns encoding information about model components

covariate_info returns a list with names continuous and categorical, with information about both continuous and categorical covariates

component_names returns a character vector with component names

get_model for lgpfit objects returns an lgpmodel

is_f_sampled returns a logical value

get_stanfit returns a stanfit (rstan)

postproc applies postprocessing and returns an updated lgpfit

clear_postproc removes postprocessing information and returns an updated lgpfit

num_paramsets, num_evalpoints and num_components return an integer

Functions

parameter_info: Get parameter information (priors etc.).

component_info: Get component information.

covariate_info: Get covariate information.

component_names: Get component names.

get_model: Get lgpmodel object.

is_f_sampled: Determine if signal f is sampled or marginalized.

get_stanfit: Extract stanfit object.

postproc: Perform postprocessing.

contains_postproc: Determine if object contains postprocessing information.

clear_postproc: Clear postprocessing information (to reduce size of object).

num_paramsets: Get number of parameter sets.

num_evalpoints: Get number of points where posterior is evaluated.

num_components: Get number of model components.

See also